refactor: 供应方需求方,查看全部数据,只能修改删除自己创建的数据

+++ app/controllers/api/small_routine/buyers_controller.rb
@@ -5,8 +5,12 @@ class Api::SmallRoutine::BuyersController < Api::SmallRoutine::BaseController
   def index
     page = params[:page] || 1
     per = params[:per] || 20
-    record = Buyer.where(user_id: @current_wx_user.user_id).order("created_at desc")
-    record = record.ransack(q_params).result(distinct: true)
+    type = params[:type] 
+    if type == "all"
+      record = Buyer.all.order("created_at desc")
+    else
+      record = Buyer.where(user_id: @current_wx_user.id).order("created_at desc")
+    end
     @record = Kaminari.paginate_array(record).page(page).per(per)
   end
 
@@ -29,6 +33,10 @@ class Api::SmallRoutine::BuyersController < Api::SmallRoutine::BaseController
   end
 
   def update
+    if @current_wx_user.user_id != @record.user_id
+      return render json: {status: 400, message: "不是本人创建,无法修改"}
+    end
+    
     if @record.update!(buyer_params)
       result = [200, '修改成功']
     else
@@ -38,6 +46,10 @@ class Api::SmallRoutine::BuyersController < Api::SmallRoutine::BaseController
   end
 
   def destroy
+    if @current_wx_user.user_id != @record.user_id
+      return render json: {status: 400, message: "不是本人创建,无法删除"}
+    end
+    
     if @record.destroy
       result = [200, '删除成功']
     else
